pyproject.toml
[project]
name = "lutron-mlactionsapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-identity==1.15.0",
    "azure-keyvault-secrets==4.7.0",
    "crewai==0.121.0",
    "en-core-web-sm",
    "fastapi==0.115.6",
    "httpx==0.28.1",
    "huggingface-hub>=0.36.0",
    "keybert==0.7.0",
    "langchain-openai==0.2.14",
    "nltk==3.9.1",
    "numpy==1.24.4",
    "openai==1.75.0",
    "pydantic==2.11.9",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.20",
    "scikit-learn>=1.7.2",
    "spacy==3.7.2",
    "tiktoken==0.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "uvicorn[standard]==0.24.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl" }
